DataRetrieval.jl Documentation

DataRetrieval.jl Documentation

Introduction

Installation

API Reference

Introduction

DataRetrieval.jl is a Julia package for retrieving data USGS water data and is based off of the popular R package of the same name. Both R and Python versions of data retrieval are available: R, Python.

Installation

Currently only the development version of DataRetrieval.jl is available. To install the development version, you can use the Julia package manager and point to the git repository:

] add https://code.usgs.gov/wma/iow/computational-tools/DataRetrieval.jl.git

When developing in Julia, we recommend using the Revise package so that you do not need to keep re-compiling the package every time you make a change.

API Reference

DataRetrieval._getbaseURL — Method

_getbaseURL(service)

Function to get the base URL string based on the service being queried.

DataRetrieval._qwdata_url — Method

_qwdata_url(url)

Function to add additional information for the qwdata query.

DataRetrieval._readRDB — Method

_readRDB(response)

Private function to parse the API response from an RDB query.

Consider padding site number rather than allowing it to be an integer that loses the preceeding 0s. R version returns site number as string w/ the full 8 digits included preceding 0s. Need to resolve the column names as well, should be able to read them from the header information somehow. R has additional functionality of being able to specify a timezone when data is that granular, could add this too.

DataRetrieval._readWaterML — Method

_readWaterML(response)

Private function to parse the response body buffer object from a WaterML query.

DataRetrieval._reformat_format — Method

_reformat_format(format, service)

Function to reformat the format string based on the service queried

DataRetrieval.constructNWISURL — Method

constructNWISURL(siteNumbers; parameterCd="00060",
    startDate="", endDate="", service="site", statCd="00003",
    format="xml", expanded=True, ratingType="base", statReportType="daily",
    statType="mean")

Construct a URL to be used to make an API query to the NWIS service.

DataRetrieval.readNWIS — Method

readNWIS(obs_url)

Function to take an NWIS url (typically constructed using the constructNWISURL() function) and return the associated data.