API Reference

Documentation for DataRetrieval.jl's public functions.

Index

DataRetrieval.constructNWISURL

DataRetrieval.readNWIS

DataRetrieval.readNWISdv

DataRetrieval.readNWISiv

DataRetrieval.readNWISpCode

DataRetrieval.readNWISqw

DataRetrieval.readNWISqwdata

DataRetrieval.readNWISunit

DataRetrieval.readNWISuv

Reference

DataRetrieval.readNWISdv — Function

readNWISdv(siteNumbers, parameterCd;
           startDate="", endDate="", statCd="00003")

Function to obtain daily value data from the NWIS web service.

Examples

julia> df, response = readNWISdv("01646500", "00060",
                                 startDate="2010-10-01", endDate="2010-10-01");

julia> df  # df contains the formatted data as a DataFrame
1×5 DataFrame
 Row │ agency_cd  site_no   datetime    68478_00060_00003  68478_00060_00003_c ⋯
     │ String7    String15  String15    String7            String3             ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ USGS       01646500  2010-10-01  13100              A                   ⋯
                                                                1 column omitted
julia> typeof(response)  # response is the unmodified HTTP GET response object
HTTP.Messages.Response

DataRetrieval.readNWISpCode — Function

readNWISpCode(parameterCd)

Function to obtain parameter code information from the NWIS web service.

Examples

julia> df, response = readNWISpCode("00060");

julia> df  # df contains the formatted data as a DataFrame
1×13 DataFrame
 Row │ parameter_cd  group     parm_nm                           epa_equivalen ⋯
     │ String7       String15  String                            String15      ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ 00060         Physical  Discharge, cubic feet per second  Not checked   ⋯
                                                              10 columns omitted

julia> typeof(response)  # response is the unmodified HTTP GET response object
HTTP.Messages.Response

DataRetrieval.readNWISqw — Function

readNWISqw(siteNumbers;
           startDate="", endDate="", expanded=true)

Function to obtain water quality data from the NWIS web service.

DataRetrieval.readNWISqwdata — Function

readNWISqwdata(siteNumbers;
               startDate="", endDate="", expanded=true)

Alias to readNWISqw().

DataRetrieval.readNWISunit — Function

readNWISunit(siteNumbers, parameterCd; startDate="", endDate="")

Function to obtain instantaneous value data from the NWIS web service.

Examples

julia> df, response = readNWISunit("01646500", "00060",
                                   startDate="2022-12-29",
                                   endDate="2022-12-29");

julia> first(df)  # df contains the formatted data as a DataFrame
DataFrameRow
 Row │ agency_cd  site_no   datetime          tz_cd    69928_00060  69928_0006 ⋯
     │ String7    String15  String31          String3  String7      String3    ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ USGS       01646500  2022-12-29 00:00  EST      12700        P          ⋯
                                                                1 column omitted

julia> typeof(response)  # response is the unmodified HTTP GET response object
HTTP.Messages.Response

DataRetrieval.readNWISuv — Function

readNWISuv(siteNumbers, parameterCd; startDate="", endDate="")

Alias for readNWISunit().

DataRetrieval.readNWISiv — Function

readNWISiv(siteNumbers, parameterCd; startDate="", endDate="")

Alias for readNWISunit().

DataRetrieval.constructNWISURL — Function

constructNWISURL(siteNumbers; parameterCd="00060",
    startDate="", endDate="", service="site", statCd="00003",
    format="xml", expanded=True, ratingType="base", statReportType="daily",
    statType="mean")

Construct a URL to be used to make an API query to the NWIS service.

DataRetrieval.readNWIS — Function

readNWIS(obs_url)

Function to take an NWIS url (typically constructed using the constructNWISURL() function) and return the associated data.